+++ osu.Desktop/OsuGameDesktop.cs
@@ -41,7 +41,7 @@ namespace osu.Desktop
             }
             catch (Exception)
             {
-                Logger.Log("Could not find a stable install", LoggingTarget.Runtime, LogLevel.Important);
+                Logger.Log("找不到osu!stable的安装位置!", LoggingTarget.Runtime, LogLevel.Important);
             }
 
             return null;