+++ osu.Game/Overlays/Mods/ModSelectOverlay.cs
@@ -127,7 +127,7 @@ namespace osu.Game.Overlays.Mods
                                         {
                                             new OsuSpriteText
                                             {
-                                                Text = @"游戏Mod",
+                                                Text = @"额外玩法",
                                                 Font = OsuFont.GetFont(size: 22, weight: FontWeight.Bold),
                                                 Shadow = true,
                                                 Margin = new MarginPadding
@@ -143,7 +143,7 @@ namespace osu.Game.Overlays.Mods
                                             {
                                                 RelativeSizeAxes = Axes.X,
                                                 AutoSizeAxes = Axes.Y,
-                                                Text = "可以让你以不同的方式来进行游戏。有些 Mod 对你的游戏分数会有影响\n还有一些仅供娱乐。",
+                                                Text = "他们可以让你以不同的方式来进行游戏。有些玩法对你的游戏分数会有影响\n还有一些仅供娱乐。",
                                             },
                                         },
                                     },
@@ -219,7 +219,7 @@ namespace osu.Game.Overlays.Mods
                                             DeselectAllButton = new TriangleButton
                                             {
                                                 Width = 180,
-                                                Text = "重置Mods",
+                                                Text = "恢复默认玩法",
                                                 Action = DeselectAll,
                                                 Margin = new MarginPadding
                                                 {